Restructured everything. Makefile and Tests need fix.
