--- HGWOSCA/Makefile
@@ -1,43 +1,0 @@
-# ==================================================================================== #
-# = Executable
-# ==================================================================================== #
-
-BIN = tests
-
-# ==================================================================================== #
-# = Compiler settings
-# ==================================================================================== #
-
-CC      = gcc
-CFLAGS += -O3 -fno-tree-vectorize -std=c11
-LIBS   += -lm -lcriterion
-
-# ==================================================================================== #
-# = Object / Assembly Files
-# ==================================================================================== #
-
-%.o : %.c
-	$(CC) $(CFLAGS) -c $< -o $@
-
-%.s : %.c
-	$(CC) $(CFLAGS) -S $< -o $@
-
-SRCS=$(wildcard *.c)
-
-OBJS=$(SRCS:.c=.o)
-
-# ==================================================================================== #
-# = Targets
-# ==================================================================================== #
-
-ASMS=$(SRCS:.c=.s)
-
-all: $(OBJS) $(ASMS)
-	$(CC) $(OBJS) -o $(BIN) $(LIBS)
-
-test:
-	./tests
-
-clean:
-	rm -rf *.o *.txt *.s
-	rm -rf $(BIN)
